requirements.txt
streamlit==1.43.2
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.25.0
SQLAlchemy==2.0.40
cryptography==44.0.2
psycopg2-binary==2.9.11
python-jose==3.3.0
python-multipart==0.0.9
redis==5.2.1
bcrypt==4.2.1
joblib==1.4.2
python-dotenv==1.1.1
pytest==8.3.5
pytest-cov==6.1.0
ruff==0.11.8
